Chore : Input, Output 분리
