+++ src/main/java/io/CalculatorScannerInput.java
@@ -5,36 +5,15 @@ import exception.ExceptionMessage;
 import java.util.Scanner;
 import java.util.regex.Pattern;
 
-public class CalculatorConsole implements CalculatorOutput, CalculatorInput {
-    private static final String BUTTON_INIT_MESSAGE = "1. 조회\n2. 계산";
-    private static final String BUTTON_CHOICE_MESSAGE = "선택 : ";
-    private static final String RECORD_MESSAGE = "%s = %d";
+public class CalculatorScannerInput implements CalculatorInput{
     private static final Pattern BUTTON_PATTERN = Pattern.compile("^[1-2]$");
     private static final Pattern CALCULATION_FORMULA_PATTERN = Pattern.compile("^[0-9]+(\\s[+\\-*/]\\s[0-9]+)+$");
-
-    private final Scanner sc = new Scanner(System.in);
-
-    @Override
-    public void printButtonMessage() {
-        System.out.println(BUTTON_INIT_MESSAGE);
-        System.out.print(BUTTON_CHOICE_MESSAGE);
-    }
-
-    @Override
-    public void printResult(int result) {
-        System.out.println(result);
-    }
-
-    @Override
-    public void printRecords(String record, int result) {
-        System.out.printf(RECORD_MESSAGE, record, result);
-    }
+    private static final Scanner sc = new Scanner(System.in);
 
     @Override
     public int inputButton() {
-        String input = sc.next();
+        String input = sc.nextLine();
         validateButton(input);
-
         return Integer.parseInt(input);
     }
 
@@ -48,7 +27,6 @@ public class CalculatorConsole implements CalculatorOutput, CalculatorInput {
     public String inputCalculationFormula() {
         String input = sc.nextLine();
         validateCalculateFormula(input);
-
         return input;
     }
